\name{tcltk.package}
\alias{tcltk.package}
\alias{tcltk}
\docType{package}
\title{
Tcl/Tk Interface
}
\description{
Interface and language bindings to Tcl/Tk GUI elements.
}
\details{
This package provides access to the platform-independent
Tcl/Tk GUI elements.  See \link{TkWidgets} for a list of
supported widgets, \link{TkWidgetcmds} for commands to work with them,
and references in those files for more.  

The Tcl/Tk documentation is in
#ifdef windows
\code{R_HOME/Tcl/doc}.
#endif
#ifdef unix
the system man pages.
#endif.

For a complete list of functions, use \code{ls("package:tcltk")}.

}
\author{
R Development Core Team

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
